pyproject.toml
[tool.poetry]
name = "nav-assist-agent"
version = "0.1.0"
description = ""
authors = ["Arnav Varyani <arnavvaryani@gmail.com>"]
readme = "README.md"
packages = [
    {include = "components"},
    {include = "services"},
    {include = "metrics"},
]

[tool.poetry.dependencies]
python = "^3.11"
browser-use = "^0.1.41"
nest-asyncio = "^1.6.0"
streamlit = "^1.44.1"
playwright = "^1.51.0"
python-dotenv = "^1.1.0"
matplotlib = "^3.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
